requirements.txt
numpy>=1.24
Pillow>=10.0
requests>=2.28
pyproj>=3.5
contourpy>=1.1
tifffile>=2023.1
pyinstaller>=6.0
